# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DetectRequest {
  /** Source code to analyze */
  code: string;
  /** Optional filename hint (e.g. "main.py") */
  filename?: string;
}

export interface LanguageCandidate {
  language: string;
  confidence: number;
  signals: string[];
}

/**
 * high >= 0.92, medium 0.70-0.91, low < 0.70
 */
export type DetectResponseConfidenceLevel =
  (typeof DetectResponseConfidenceLevel)[keyof typeof DetectResponseConfidenceLevel];

export const DetectResponseConfidenceLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface DetectResponse {
  /** Primary detected language */
  detected: string;
  /** Confidence score (0-1) */
  confidence: number;
  /** high >= 0.92, medium 0.70-0.91, low < 0.70 */
  confidenceLevel: DetectResponseConfidenceLevel;
  candidates: LanguageCandidate[];
  /** Detection signals used */
  signals: string[];
  /** Time taken for detection in milliseconds */
  detectionMs: number;
}

export interface CompileRequest {
  /** Source code to compile/run */
  code: string;
  /** Optional filename hint */
  filename?: string;
  /** Override language detection */
  language?: string;
  /** Additional arguments to pass to the compiler */
  args?: string[];
}

export type CompileResponseConfidenceLevel =
  (typeof CompileResponseConfidenceLevel)[keyof typeof CompileResponseConfidenceLevel];

export const CompileResponseConfidenceLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface CompileResponse {
  submissionId: number;
  detected: string;
  confidence: number;
  confidenceLevel: CompileResponseConfidenceLevel;
  stdout: string;
  stderr: string;
  exitCode: number;
  success: boolean;
  compilationMs: number;
  toolchainAvailable: boolean;
  /** Install instructions if toolchain not available */
  installHint?: string;
}

export interface Submission {
  id: number;
  code: string;
  filename?: string | null;
  detectedLanguage: string;
  confidence: number;
  confidenceLevel: string;
  stdout: string;
  stderr: string;
  exitCode: number;
  success: boolean;
  compilationMs: number;
  createdAt: string;
}

export interface SubmissionListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ToolchainEntry {
  language: string;
  available: boolean;
  command: string;
  version?: string | null;
  installHint: string;
}

export interface ToolchainListResponse {
  toolchains: ToolchainEntry[];
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type ListSubmissionsParams = {
  /**
   * Page number (1-based)
   */
  page?: number;
  /**
   * Items per page
   */
  limit?: number;
  /**
   * Filter by detected language
   */
  language?: string;
};