# lib/api-zod/src/generated/types/compileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface CompileRequest {
  /** Source code to compile/run */
  code: string;
  /** Optional filename hint */
  filename?: string;
  /** Override language detection */
  language?: string;
  /** Additional arguments to pass to the compiler */
  args?: string[];
}