# lib/api-zod/src/generated/types/compileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { CompileResponseConfidenceLevel } from "./compileResponseConfidenceLevel";

export interface CompileResponse {
  submissionId: number;
  detected: string;
  confidence: number;
  confidenceLevel: CompileResponseConfidenceLevel;
  stdout: string;
  stderr: string;
  exitCode: number;
  success: boolean;
  compilationMs: number;
  toolchainAvailable: boolean;
  /** Install instructions if toolchain not available */
  installHint?: string;
}