# lib/api-zod/src/generated/types/detectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectRequest {
  /** Source code to analyze */
  code: string;
  /** Optional filename hint (e.g. "main.py") */
  filename?: string;
}