# lib/api-zod/src/generated/types/detectResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectResponseConfidenceLevel } from "./detectResponseConfidenceLevel";
import type { LanguageCandidate } from "./languageCandidate";

export interface DetectResponse {
  /** Primary detected language */
  detected: string;
  /** Confidence score (0-1) */
  confidence: number;
  /** high >= 0.92, medium 0.70-0.91, low < 0.70 */
  confidenceLevel: DetectResponseConfidenceLevel;
  candidates: LanguageCandidate[];
  /** Detection signals used */
  signals: string[];
  /** Time taken for detection in milliseconds */
  detectionMs: number;
}