# lib/api-zod/src/generated/types/languageCandidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface LanguageCandidate {
  language: string;
  confidence: number;
  signals: string[];
}