# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface Submission {
  id: number;
  code: string;
  filename?: string | null;
  detectedLanguage: string;
  confidence: number;
  confidenceLevel: string;
  stdout: string;
  stderr: string;
  exitCode: number;
  success: boolean;
  compilationMs: number;
  createdAt: Date;
}