# lib/api-zod/src/generated/types/toolchainListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polyglot Auto-Detecting Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolchainEntry } from "./toolchainEntry";

export interface ToolchainListResponse {
  toolchains: ToolchainEntry[];
}